feat: add settings `overrideReplacesDocs`, `augmentsExtendsReplacesDocs`, `implementsReplacesDocs` and apply to  `require-param`, `require-description`, `require-example`, and `require-returns` for exempting all of these from errors in the case of the presence of `@override`, `@augments`/`@extends`, or `@implements`, respectively.
This commit also adds the following new behavior: with `require-description`, `inheritdoc` will avoid erring.

This commit also causes the following non-breaking changes:

With `require-example` and `require-returns`, `@override` may now err if the setting `overrideReplacesDocs` is set to false
With `require-returns`, `@implements` may now err if the setting `implementsReplacesDocs` is set to false
With `require-example`, `require-returns`, and `require-description`, `@augments` or `@extends` may now avoid erring if the setting `augmentsExtendsReplacesDocs` is set to true

Deprecating `settings.jsdoc.allowOverrideWithoutParam`, `settings.jsdoc.allowImplementsWithoutParam`, and `settings.jsdoc.allowAugmentsExtendsWithoutParam` in favor of the more tag-independent settings.