feat(`getJSDocComment`): ignore line comment
Often it is required to disable linting in-between the JSDoc and the
code. This PR allows that to happen.

Without the following change the `jsdoc/require-jsdoc` rule needs to be
constantly disabled.

Consider the following example:

```ts
/**
 * A user.
 * @interface User
 */
interface User {
  /**
   * The users age.
   * @type {number}
   * @memberof User
   */
  age: number;

  /**
   * The users name.
   * @type {string}
   * @memberof User
   */
  name: string;
}

/**
 * Determine if the provided value is a `User`;
 * @param {*} user The value to check.
 * @returns {boolean} Whether or not the provided value is a `User`.
 */
// Allow the Function to check any input.
// tslint:disable-next-line: no-any
function isUser(user: any): user is User {
  if (user == null) {
      return false;
  }
  const { age, name } = user;
  if (typeof age !== `number`) {
      return false;
  }
  if (typeof name !== `string`) {
      return false;
  }
  return true;
}
```

In the above example, using `unknown` wouldn't cut it. This is because
`unknown` cannot be destructured. Eg: `const { age, name } = user;`.

This PR also updates the checkExamples.js assertion so as not to require
an absolute path to @typescript-eslint/parser.

This absolute path would mean constant calls to npm run create-readme
because of the different paths that developers would have the repo
checked out at.