feat(`validateDescription`): tighten constraint
Currently it is possible to experience a false positive with the
`jsdoc/require-description-complete-sentence` rule.

The false positive is caused by starting a new line with a capital
letter when the previous line does not end with a period. Normally, this
is the correct behaviour, however, English sentences are more nuanced
than that.

The problem arises when using a word that starts with a capital letter
that just so happens to be at the start of a line. For instance, `I`, or
a programmatic word like `XMLHttpRequest`.

This PR tightens the constraint of the `isNewLinePrecededByAPeriod`
Function by checking for a lowercase letter immediately after an
uppercase letter.

This will still produce false positives if the new line starts with
`UserInterface` for example.