fix(`newline-afer-description`): when finding last description line, don't look beyond the length of the description.
The previous check for the last description line was redundant and problematic.
The search for the last line was mistakenly checking for the last match within the full comment, so if the last line of the
description were duplicated later, that position was used, and since that position couldn't be find within the
`comment-parser` description, there would be an error. It was also mistakenly stripping carriage returns in part
of the comparison when there could be carriage returns in the other part.