feat(`require-description`, `require-example`, `require-param`, `require-returns`): cause `exemptedBy` to overwrite `inheritdoc` exempting; closes #510
BREAKING CHANGE:

While still defaulting to having `inheritdoc` exempt rules, for the impacted rules,
if one has defined an `exemptedBy` options array one must now explicitly add
back `"inheritdoc"` to get the exemption for that tag as well. This change allows
users to omit `inheritdoc` from the array so as to cause its presence not to
trigger exemption of the rule.